Add NonZero<usize> hack (remove once NonZero is stabilized)
This hack saves quite a bit of memory. It uses the fact that Rust optimizes an Option<&T> to be the same size as a &T (8 bytes), since Rust knows internally that pointers can never be 0x0 / NULL.

The `NonZeroUsizeHack` provides a safe interface to this by adding 1 to the index on creation and removing 1 on retrieval - the IDs are still the same, but now the `NodeId` is only 8 bytes large instead of 16 bytes.