Refactor the DisplayList.rectangles so that we have the NodeId / DOM information in the layout step
Now the `get_wh_for_rectangle()` actually works! It outputs 400, which is correct - the rectangle has a width of 200, but a min-width of 400, so the min-width takes precedence.

For traversing the DOM, we need information about parent / ancestor, etc. in the layout step.

Added LinearIterator for iterating the Arena linearly (without DOM information, just iterate the inner Vec).