Migrated from Arc<Mutex<AppState<T>>> to Arc<Mutex<T>>
For implementing the task system we can't send the whole AppState across threads.
The AppState contains windows and GlContexts, which aren't thread-safe to access.
So in the task system we can only send an Arc<Mutex<T>> and the rest of the
operations happen on the main thread anyway. So it doesn't make sense to use
an Arc<Mutex<AppState>> when we can't even send the AppState across threads.