Make dynamic CSS properties do something, add window state diffing, prepare asynchronous window updates
Currently, the design is to let the user set properties on the `AppState`, and then, after the redraw, check what windows need updating / if there are new windows and set the new CSS properties.

For this, we want to decouple the actual window from the state of the window and the dynamic CSS properties from the window. So we expose a Vec<WindowState> on the AppState, which the user can freely modify during the callbacks. Afterwards, we diff the state and check for changes.

This is also a preparation for CSS hover events (which are going to be implemented using special dynamic CSS IDs).