Added FakeCss and FakeWindow, allow user to manipulate the state of the window
This means that dynamic properties in CSS now work as intended, as does updating / diffing the window state for properties.

There is currently a huge bug in the timing - the overridden CSS values are not updated when the change occurs. See the notes in the app.rs file.

NOTE: We should really migrate from a Vec<WindowId> to a HashMap, in order to make addition / removal of windows not dependent on their position in the Vec.