Initial implementation of text caching implemented
This version of text caching is absolutely awful, since it simply
clones large strings every time the text is resized or scrolled or
whatever. However it's just a first implementation and there are
bigger issues to worry about right now.

Another danger is that you could lose a text ID, thereby "leaking"
the text string. This should be revisited, so that a proper error
is returned.

In general, the text string itself should not be stored with the
layout of the glyphs together - because every time you'd want to
change the glyphs, this would mean that you'd have to change the
text, too, which isn't the case.

This commit also adds image caching (so that you can either push
images via an image ID or a CSS identifier (aka a String)). Doing
it via the String requires 1 extra lookup, but is more convenient.