Fixed the DOM add_child() and add_sibling() functions
See: https://gist.github.com/fschutt/4b3bd9a2654b548a6eb0b6a8623bdc8a
for the original algorithm in Python. Essentially we copy the child /
sibling nodes to the self.nodes and just fix up the node IDs
respectively.

While this is highly unsafe, it is performant in practice. I think
that we can also remove the RefCell now, since that was only needed
in the broken implementation, which relied on cloning the DOM
node-for-node.