Fixed order of input handling / redrawing commands
Currently I don't know why the order has to be this way: First check for input, then do the drawing. In theory they should be independent from each other, but in another order the display doesn't get swapped correctly (?).

I.e. I moved the if frame_event_info.should_hittest { } above the should_swap check.

Also migrated to a WindowCloseEvent to not rely on a boolean anymore (more expressive) + use a Vec<On> to indicate no events instead of an Option<Vec<On>> - eliminates branching and Vec::new() does not allocate, so in the (likely) case that we have no input events to react to, it'll be a bit faster.