Added FromIterator<Item=NodeData | Dom> for Dom
You can now write:

  let dom = (0..5).map(|i|
     NodeData::new(NodeType::Label(format!("{}", i)), .. Default::default())
  ).collect();

or:

  let dom = (0..5).map(|i| NodeType::Label(i.to_string())).collect()

Both will generate the same DOM, but the first method allows
you to customize the ID or class of the iterated item.

This enables very terse code when working with lists. The
resulting DOM has all resulting element as siblings, meaning
they have no parent. Also added checks if the number of DOM
nodes is 0 (if no nodes are present in the DOM).