Initial architecture for CSS to layout constraints
This represents the change from hard-coded constraints to transitioning to the cassowary constraint solver.

Right now it should be possible to transform the CSS layout parameters into layout constraints - however, the initial test doesn't work yet.

DisplayRect has been renamed to RectConstraintVariables (since it contains the variable IDs for the rectangle), to not be confused with DisplayRectangle.

Right now if the DOM changes, the memory for the constraints is "leaked", i.e. the constraints aren't removed. This will be fixed in future versions.