Added CSS hot reload + "fixed" memory leak in constraint solver
For now, all constraints are simply removed after the frame, so that we don't create excess constraints in the solver

Hot reloading CSS files is now possible, but only in debug mode (to prevent accidentally releasing an app without embedding the CSS). Because of this, the CSS simply gets reloaded every 500ms, edited or not.

box-shadow has a slight bug regarding layout, as well as linear-gradient, but the important part is that it all mostly works now.

What's missing from the proper layout is absolute / relative layout and multi-window handling, but then the first version of this library should be done.