"Fixed" the OpenGL crashing issue by simply not drawing invalid textures
This isn't a clean solution, but on the other hand, we won't have
any crashing issues with panics anymore - if a texture ID isn't found
we simply return ExternalImageSource::Invalid

The issue seems to be that cleanup_unused_opengl_textures is called
while the window is minimized or at least before it is restored.
However, winit doesn't have a clean API to check if the window was
minimized, so adding this would overcomplicate the main loop. Not
drawing anything is way simpler.