Preliminary callback-based OpenGL rendering
This enables users to know the bounds of the rectangle
they are rendering into. For example, if we want the
OpenGL texture to be inside of a rectangle, we need to
know the width and the height of the texture, so we can
draw the thing we want to draw properly.

For this, we stop the rendering process, lock the users
data, and invoke a user-defined callback (the
GlTextureCallback) with the width and the height that the
layout solver gave us.

This way the user knows the width and height of the
component it should render. Right now, the layout doesn't
work very stable - the rectangle can jitter around, but
at least it draws properly.