Remove menu system (preparation for run_in_thread)
Menu handling never worked, since Win32 (or winit, not sure) locks
up the application if an action, such as AppendMenuW on Windows
is not run from the thread that the window was created from.

Winit has a special API, which lets the user run a Fn closure
inside a special thread. This could be used to make a menu
into a "Widget" rather than a special-cased API.

The menu system never worked, so this commit removes it.