Fixed off-by-one error in line breaking function in text_layout.rs
The weird spacing prior to this commit was caused because we were
using an exclusive range on the characters, i.e. LineBreak(4) would
take characters 0..4 which **excludes** the 4th character itself.

Now we do 0..=4, which includes the character, but we also have to
be aware that the next line starts at 5.. (so the character count
has to be increased by 1 for the next line).