Table render demo doesn't crash anymore
In the push_iframe function, we have to push
display_list.rectangles not push the parent display list (bug!).

Also, temporarily removed the reparenting of the DOM in the
UiState::from_dom function - since the .add_child() clears
the UiDescription arena via the Rc it has, it leads to a crash.

Also added a nice .print_tree() for debugging the DOM. Right
now the demo still doesn't display anything (seems to hang
in the Changeset calculation function), but at least it doesn't
crash anymore.