Wrote unit-tests for flexbox layout
Fixed a small bug with != Some(LayoutPosition::Absolute) instead
of == Some(LayoutPosition::Absolute). Right now the space is evenly
distributed, as expected: the paddings are respected (haven't tested
margins yet). However, the max-width isn't respected, likely because
that's not part of flexbox. That may need a seperate pass after the
initial flex solver in order to work correctly.