Removed flex-shrink for now, layout step is now constraint-aware
I don't know how flex-shrink works when non-flexbox constraints
such as max-width / width / min-width come into play. So flex-shrink
is simply removed for now (but still left in the CSS parser).

For now, the flex-grow distributes across the remaining, flexible
items, not all child-items. Not sure if that is browser-compatible,
but it makes the most sense.