Removed add_sibling, fixed table layout bugs
add_sibling() and with_sibling() are removed, because they are very
hard to reason about, since they act like an add_child() on the parent,
which is simply weird in practice, since you wouldn't expect a top-down
tree to be able to affect its parent. This led to the current bugs in
the table_view.rs - which are now fixed, but add_sibling() has to go,
it's just too confusing.

This also means that the root node won't have the ability to have
siblings - however, this is simply worked around with a root DOM node
with children instead of one sibling.

This also removes the need to re-allocate the DOM in the UiDescription,
since we now know that the DOM returned by the user has exactly one
root component (because we disallowed siblings).

Other than that, the bubbling-to-the-parent function is now only run
once and is O(n) instead of O(n ^ 2).