Removed slow cassowary solver, table demo working
Cassowary has a O(n³) algorithm for layout solving, which creates
layout times of several seconds! The table demo was unable to run
previously, because the solver needed 12 seconds just to solve
the layout.

With a custom flexbox solver, we are now able to render the table
in roughly 3ms - 5ms (in release mode, and this is a large table).

Therefore we don't need the original layout solver anymore. Right
now there are a few bugs regarding clipping, but overall it works
quite well. The table demo has currently some weird layout problems
with the row numbers, but at least it runs! This means that it's
now possible to display infinite lists or very large datastructures.