Fix list demo and make text input less crash-happy
If a user allocated a TextInputState on the heap, previously the
text input would simply crash. Now it just emits a "On::KeyDown"
signal to signify that it can receive KeyDown events, so that a
user can call TextInput::on_key_down() from outside the component.

The fix was to always submit a rectangle to the display list,
independent of the background color. Otherwise, the hit-testing
function couldn't find the rectangle to test against.