Fixed image filling aspect ratio + box shadow order
The box shadow has to be pushed after the content, but before the
scrollbar.

Since the image width / height gets respected in the layout pass,
there is no need for adjusting the width and height of the image
later on.

This will however probably lead to bad results with CSS backgrounds
(? - untested).