Minor edits to README.md
"Rusts" --> "Rust's" (in the "Programming Model" section)
Edited the first paragraph for more clarity. I think you were saying that Azul supports development of desktop apps that were written in Rust AND were supported by the Mozilla WebRender rendering engine, but the original sentence was unclear about that. I rewrote that part and then wrote about Azul using a CSS/DOM model in a separate sentence. It's to be clear that Azul uses the model, not the desktop apps (if that was what you meant).