Fix missing features + rusttype dependency bullshit
Rusttype upgraded stb_truetype to 0.2.5 (a minor version bump) while also upgrading
to Rust 2018, which caused a major breakage, because now this project is
incompatible with Rust 1.30.

Slight rant incoming: I am increasingly getting sick of rusttype, at least for this
project. Yes, it's pure-Rust, but in practical terms that doesn't mean shit. Back in
February, rusttype managed to break azuls entire dependencies by yanking rusttype
from crates.io and only releasing a major-version-breaking bugfix (this type of
yanking also breaks cargo in an npm-leftpad-style fashion, and while I reported this
problem to the cargo team, it got completely ignored, yay "we hear you").

Every time I've had build problems in the past, it was mostly rusttype that broke
the build. I can't rely on such an unstable crate in my dependencies, for fucks
sake, get your shit together.

Add to that that rusttypes API is pretty horrible (what metric is "rusttype::Scale"?
- oh yeah, it's just "scale", not "millimeter" or "point" or anything meaningful),
rusttype doesn't support OTF or PostScript fonts, just the most bare-bones TTF
fonts, the kerning is buggy as shit, meaning that half the time it doesn't work
("fi" isn't getting kerned, even in fonts that do support it). Plus the constant
version breaking and we have to ship freetype anyway, as it's a requirement for
webrender, so it just adds dependencies while limiting the use of fonts to just TTF
ones.

So yeah, I'll start to remove rusttype from my dependencies in the future and switch
to freetype. I'm sick of this shit.