Implement no-repeat, repeat, repeat-x, and repeat-y for background-repeat
This was implemented by using LayoutPrimitiveInfo's clip_rect functionality.
If `no-repeat` is specified in CSS, the rect is clipped to the size of the image on both axis.
If `repeat-x` is specified, the rect's height is clipped to the image height.
If `repeat-y` is specified, the rect's width is clipped to the image width.`