Make CallbackInfo::set_focus() functional
Users can now set the focus of the application on the next frame
using either a CSS Path (similar to JS) or a NodeId, as well as
un-focus certain

What is currently not done is that FocusLeft / FocusReceived events
are fired (in the case that the focus is set programmatically
instead of the application user doing that). This is an area that
will need to be improved in the future (likely to be implemented
when tab-navigation will be implemented).