Refactored image / font API to be generated by a macro
Instead of creating the API manually or using Deref, it's
easier to create the API via a macro, so that refactoring
takes less time.

Also refactored CssImageId and ImageId to be separate from
each other - this breaks the current image API, but is a
preparation for ImageSource / FontSource.

Removed add_custom_task in favor of add_task - the Task<T>
already takes care of checking the T: Send + 'static, so
we don't have to create two separate functions for the same
thing.