Fixed font + image GC to work correctly
The fonts and images have to be split into "current" and "pending"
transaction, otherwise, they will constantly interact with eachother,
which leads to infinite add-remove-add-remove cycles.

Instead of going through all fonts that are currently registered, we
only go through the fonts that have been added on frame n and compare
them with the fonts that were present in frame (n-1) and add / remove
fonts accordingly.

The only downside is that on startup (on the first frame), the fonts
are added, deleted, then added again. However, this should not have
any visual consequences (since the fonts are added before the DOM is
drawn at all).