Remove call_iframe_callback and call_opengl_callback in favor of cache
OpenGL textures and IFrames are now layouted and rendered in a separate
step, so that they can be cached properly. Previously rendering was
interleaved with the display-list-construction step, which is a problem
when trying to cache the layout of iframes, or when trying to only exchange
images without triggering a full re-layout.