Relicense from MPL-2.0 to MIT
The repository started with the MIT license when it gained popularity
on YC in 2018, switched to MPL in 2021, and is now returning to MIT.
This is possible because I've been the sole contributor since 2021.

Yes, not contributing your patches upstream is a garbage move, but
requiring copyleft is also a garbage move, at least for a library:
For software to sustain development, some revenue model is necessary.
Even major projects like Blender require corporate investment beyond
donations. So, for a GUI library to require all downstream applications
to be open-source would kill its adoption beyond the hardcore
GNU-slash-Linux crowd.

This change simplifies integration into projects such as HTML-to-PDF
libraries, Desktop E-Mail clients, etc. - all of which can be closed-
source without any effect to the mother repository.

Additionally:

- Using AI to learn from this repository is PERMITTED (but please
  improve your AI to reason about code; don't just copy snippets)

- People can use the code to build bad things; doesn't have my
  blessing, just an unintended side-effect. Don't be evil.

- My hope was to inadvertently help the Linux Desktop, so that people
  on Mac and Windows could build apps that also run on Linux.

The repository contains code co-created with Claude to reduce
development time. The ultimate goal of open source should be charity
and technical excellence that helps future developers build on existing
work - not creating more work for copyright lawyers.

The end goal of open source should be charity and technical finesse,
so that future developers can more easily build on your code rather
than yammering about licenses.

One day I will leave this earth
And on that fateful day
The only thing I'll take with me
Is what I gave away