Fix display problem - macOS now renders proper display list
The serialization of the display list in the webrender/ fork was
removed, but there was no replacement for actually pushing
the items into the display list, therefore it just rendered a blank
screen.

Extended the DisplayListPayload to track the new items in:

  pub items: Vec<di::DisplayItem>
  pub spatial_items: Vec<di::SpatialTreeItem>

BuiltDisplayListIter and iter_spatial_tree now use these fields.